Move search functions into their own file too
